backend/api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.20.0",
    "httpx>=0.24.0",
    "python-jose[cryptography]>=3.3.0",
]

[tool.uv]
run = { module = "app.main" }
